requirements.txt
tensorflow==1.13.1
django-extensions==2.2.9
djangorestframework==3.11.0
imageai==2.0.2
numpy==1.18.1
scipy==1.18.1
opencv-python==4.2.0.34
matplotlib==3.2.2
cycler==0.10.0
kiwisolver==1.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
h5py==1.15.0
keras==2.4.2
pyyaml==5.3.1
image==1.5.32